Fix issue with RX/TX pins of Arduino Pro Mini
In the prior commit, the WANT_REAL_SERIAL flag was accidentally left on,
making train numbers 5 and 6 unusable. In testing that, discovered that
the silk-screen on the Arduino Pro Micro is apparently wrong about the
TX0/RX1 pins.

Also, added a new station #12 on the START2 pin of the MRCS Morse Code
Buzzer Controller.

Update the documentation to reflect the current state of the code.